+++ frontend/package.json
@@ -25,6 +25,7 @@
     "react-compound-slider": "^3.3.1",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.5",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "react-window": "^1.8.6",
     "semantic-ui-css": "^2.4.1",
