+++ frontend/src/export.js
@@ -1,0 +1,3 @@
+export const Export = () => {
+  return (<h2>Export</h2>);
+}
