+++ frontend/src/routes.js
@@ -1,0 +1,7 @@
+import { Export } from "./export";
+import { Home } from "./home";
+
+export default [
+  { path: "/home", name: "Home", Component: Home },
+  { path: "/export", name: "Export", Component: Export }
+];
